chore: add debug script to simplify isolated api-docs debugging
