+++ test/scripts/apidoc/signature.spec.ts
@@ -1,12 +1,12 @@
 import { writeFileSync } from 'fs';
 import { resolve } from 'path';
-import * as TypeDoc from 'typedoc';
 import { afterAll, describe, expect, it } from 'vitest';
 import type { Method } from '../../../docs/.vitepress/components/api-docs/method';
 import { analyzeSignature } from '../../../scripts/apidoc/signature';
-import { newTypeDocApp, patchProject } from '../../../scripts/apidoc/utils';
 import { SignatureTest } from './signature.example';
 import expected_ from './signature.expected.json';
+import { loadExampleMethods } from './utils';
+
 const expected: Record<string, Method> = expected_;
 
 function prettyJson(object): string {
@@ -14,24 +14,9 @@ function prettyJson(object): string {
 }
 
 describe('signature', () => {
-  const app = newTypeDocApp();
-
-  app.bootstrap({
-    entryPoints: ['test/scripts/apidoc/signature.example.ts'],
-    tsconfig: 'test/scripts/apidoc/tsconfig.json',
-  });
-
-  const project = app.convert();
-
-  patchProject(project);
-
-  const methods: Record<string, TypeDoc.DeclarationReflection> = project
-    .getChildrenByKind(TypeDoc.ReflectionKind.Class)[0]
-    .getChildrenByKind(TypeDoc.ReflectionKind.Method)
-    .reduce((a, v) => ({ ...a, [v.name]: v }), {});
-
   describe('analyzeSignature()', () => {
     const actuals = {};
+    const methods = loadExampleMethods();
 
     it('dummy dependency to rerun the test if the example changes', () => {
       expect(new SignatureTest()).toBeTruthy();
@@ -43,6 +28,7 @@ describe('signature', () => {
 
     it.each(Object.keys(expected).sort())('%s', (name) => {
       const method = methods[name];
+      expect(method, `Method ${name} to be defined`).toBeDefined();
       const actual = analyzeSignature(method.signatures[0], null, method.name);
       actuals[name] = actual;
